requirements.txt
mysql-connector-python
reportlab
pytest
pytest-cov
python-dotenv
tabulate
PySide6
